% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.subclones.R
\name{absolute.method}
\alias{absolute.method}
\title{Absolute simplified clonal method}
\usage{
absolute.method(
  alt_reads,
  tot_reads,
  purity,
  total_cn,
  ccf_grid = seq(0, 1, by = 0.01)
)
}
\arguments{
\item{alt_reads}{Number of alternative reads for the mutation}

\item{tot_reads}{Number of total reads for the mutation}

\item{purity}{Estimated sample purity}

\item{total_cn}{Total integer copy number at the mutation's locus}

\item{ccf_grid}{Grid of ccf values to compute the distribtion (default: 0.01 - 1)}
}
\description{
\code{absolute.method} computes the cancer cell fraction, confidence intervals, and clonal or subclonal status of a given mutation
}
\details{
Using total and alternative read counts, sample purity, locus total copy number, this method computes the posterior probability distribution over given cancer cell fraction values (from 0.01 to 1). Mutations were defined as clonal if the 95% confidence interval overlapped 1, and subclonal otherwise
}
