% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.R
\name{allelic.ratio}
\alias{allelic.ratio}
\title{Calculate allelic ratio}
\usage{
allelic.ratio(vaf.dna, vaf.rna)
}
\arguments{
\item{vaf.dna}{Variant allele fraction in DNA}

\item{vaf.rna}{Variant allele fraction in RNA}
}
\description{
\code{allelic.ratio} Given DNA and RNA variant fractions, this function will
calculate the allelic ratio for a given variant.
}
\details{
The allelic ratio calculation is similar to allele specific expression, but
can also be applied to somatic mutations and copy number changed regions for
both germline and somatic variants. When the RNA fraction has increased, the
allelic ratio is given by the equation: \deqn{(vaf.rna - vaf.dna) / (1 -
vaf.dna)} and when the RNA fraction has decreased by: \deqn{(vaf.rna -
vaf.dna) / vaf.dna}

This formula provides an indirect measure of transcriptional output, but has
the advantage of not requiring copy number, purity, or ploidy information.
}
