% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.R
\name{amp.somatic}
\alias{amp.somatic}
\title{Calculate transcriptional amplification value in somatic variants.}
\usage{
amp.somatic(vaf.dna, vaf.rna, norm.cn = 2, tum.cn, purity)
}
\arguments{
\item{vaf.dna}{Variant allele fraction in DNA}

\item{vaf.rna}{Variant allele fraction in RNA}

\item{norm.cn}{Local total DNA copy number in the normal cell fraction.
(Default: 2).}

\item{tum.cn}{Local total DNA copy number in the tumour cell fraction.}

\item{purity}{Tumour purity. Numeric [0-1].}
}
\description{
\code{amp.somatic} will use DNA and RNA variant fractions, along with local
tumour copy number and purity information to calculate the amplification
value for a given somatic variant.
}
\details{
This algorithm uses tumour purity, local copy number, and the dna and rna
fractions of a given somatic variant to infer the level of transcriptional
amplification taking place.
}
