% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.R
\name{calc.amp}
\alias{calc.amp}
\title{Calculate transcriptional amplification value.}
\usage{
calc.amp(vaf.dna, vaf.rna, norm.cn = 2, tum.cn, snp.cn = NULL, purity, ID)
}
\arguments{
\item{vaf.dna}{Variant allele fraction in DNA}

\item{vaf.rna}{Variant allele fraction in RNA}

\item{norm.cn}{Local total DNA copy number in the normal cell fraction.
(Default: 2).}

\item{tum.cn}{Local total DNA copy number in the tumour cell fraction.}

\item{snp.cn}{SNP copy number. Should be equal to 0 or the major CN.}

\item{purity}{Tumour purity. Numeric [0-1].}

\item{ID}{Indicate whether the variant is somatic, \emph{0} or germline,
\emph{1}.}
}
\description{
\code{calc.amp} is a wrapper for \code{amp.somatic} and \code{amp.germline}
used to calculate amplification values across all informative variants.
}
