% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_coverage.R
\name{flag_coverage}
\alias{flag_coverage}
\title{Apply threshold cutoff function}
\usage{
flag_coverage(
  depth,
  threshold,
  test = c("greater eq", "greater", "less eq", "less")
)
}
\arguments{
\item{depth}{Value of interest}

\item{threshold}{Threshold to apply cutoff}

\item{test}{Type of test to perform.}
}
\description{
Apply threshold cutoff function
}
