% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.data.R
\name{import.data}
\alias{import.data}
\title{Import RNAmp raw data}
\usage{
import.data(
  DNA_ASE_Table,
  RNA_ASE_Table,
  Variant_Table_File,
  Segments_File = NULL,
  purity = NULL,
  ploidy = NULL,
  sample_name = NULL
)
}
\arguments{
\item{DNA_ASE_Table, RNA_ASE_Table}{Allelic depth count tables produced for DNA
and RNA sequencing from RNAmp.pl.}

\item{Segments_File}{Optional: Segments file.\cr Tab seperated file in the
format: \emph{Chromosome, Start Position, End Position, Major CN, Minor CN}.}

\item{purity}{Optional: Sample purity}

\item{ploidy}{Optional: Sample ploidy}

\item{sample_name}{Optional: Name of the sample.}

\item{Variant_Table_file}{Variant table file produced from RNamp.pl script.}
}
\description{
\code{import.data} Imports variant allelic read counts along with segmented
copy number data into an RNAmp object.
}
\details{
This function will import data DNA and RNA allele count data and merge with copy number and annotation inormation or processing by RNAmp
}
