% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.subclones.R
\name{measure.subclones}
\alias{measure.subclones}
\title{Measure subclones}
\usage{
measure.subclones(rnamp.obj)
}
\arguments{
\item{rnamp.obj}{RNAmp data object from \code{import.data}}
}
\description{
\code{measure.subclones} will take processed rnamp object and apply  a simplified method from ABSOLUTE to measure clonal and subclonal mutations
}
