% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.variants.R
\name{process.variants}
\alias{process.variants}
\title{Process RNAmp variant data}
\usage{
process.variants(rnamp.obj, dna.depth = 8, rna.depth = 30)
}
\arguments{
\item{rnamp.obj}{RNAmp data object from \code{import.data}}

\item{dna.depth}{Depth filter for DNA (Default: 8)}

\item{rna.depth}{Depth filter for RNA (Default: 30)}
}
\description{
\code{process.variants} will clean, annotate, and process variant data to
make it suitable for downstream processing.
}
\details{
After importing data with \code{import.data}, this function will \enumerate{
\item CLEAN: Set missing counts to zero, remove MT variants and reorder by
chr:pos \item MERGE: Copy number and variant data \item COMPUTE: VAF in DNA
and RNA, mutation and SNP copy numbers and allelic ratio and amp \item FLAG:
Low coverage DNA and RNA sites, diploid and copy neutral regions, false
positive heterozygous SNPs, imprinted regions, and LOH alleles \item
ANNOTATE: Informative variants for transcriptional output estimation}
Processed variants are attached to the RNAmp data object as
'variants.processed'.
}
