% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.R
\name{rna.content}
\alias{rna.content}
\title{Calculate tumour RNA content.}
\usage{
rna.content(amp, purity)
}
\arguments{
\item{amp}{Mean amplification level for a tumour.}

\item{purity}{Tumour purity. Numeric [0-1].}
}
\description{
\code{rna.content} estimates the amount of tumour derived RNA in a given
primary tumour sample based on its purity and amplification level.
}
