% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmp.R
\docType{package}
\name{RNAmp}
\alias{RNAmp}
\title{RNAmp: An R package for measuring the transcriptional output from primary
tumour sequencing data.}
\description{
The RNAmp package provides basic functions for importing and processing allele
count data to measure transcriptional output levels from variant data.
}
\section{RNAmp functions}{
 To be filled in...
}

