% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.qc.R
\name{sample.qc}
\alias{sample.qc}
\title{Get sample level QC statistics}
\usage{
sample.qc(rnamp.obj)
}
\description{
\code{sample.qc} will perform QC for a given sample.
}
\details{
After importing data with \code{import.data}, and processing with
\code{process.variants}, this function will perform sample level QC assessing
mean and median depth of coverage per variant, number of variants per sample,
number of informative variants and their coverage. See the \strong{Metrics}
section for further details.

A variant is considered as \strong{covered} if it has sufficient read depth
in both DNA and RNA as defined by \code{dna.depth} and \code{rna.depth}
thresholds in \code{process.variants}.

A variant is considered \strong{informative} if it is useful for
transcriptional output measurements. These include somatic variants, and
heterozygous germline variants found in LOH regions.
}
\section{Metrics}{

\strong{n_var}: Total number of variants \cr
\strong{n_var_cov}: Total number of variants covered . \cr
\strong{prop_var_cov}: Proportion of variants covered \cr
\strong{n_var_inf}: Number of informative variants \cr
\strong{n_var_inf_cov}: Number of covered informative variants \cr
\strong{prop_inf_cov}: Proportion of informative variants covered \cr
\strong{prop_cov_inf}: Proportion of covered variants that are informative \cr
\strong{n_som_var}: Number of somatic variants \cr
\strong{n_som_var_cov}: Number of covered somatic variants \cr
\strong{prop_som_var_cov}: Proportion of somatic variants covered \cr
\strong{n_som_var_inf}: Number of informative somatic variants covered \cr
\strong{n_som_var_inf_cov}: Number of informative somatic variants covered \cr
\strong{prop_som_var_inf_cov}: Proportion of informative somatic variants covered \cr
\strong{n_germ}: Number of germline variants \cr
\strong{n_germ_cov}: Number of covered germline variants \cr
\strong{prop_germ_cov}: Proportion of germline variants covered \cr
\strong{n_germ_inf}: Number of informative germline variants \cr
\strong{n_germ_inf_cov}: Number of informative germline variants covered \cr
\strong{prop_germ_inf_cov}: Proportion of informative germline variants covered \cr
\strong{n_germ_loh_alt}: Number of germline LOH alt variants \cr
\strong{n_germ_loh_alt_cov}: Number of germline LOH alt variants covered \cr
\strong{prop_germ_loh_alt_cov}: Proportion of germline LOH alt variants covered \cr
\strong{n_germ_loh_ref}: Number of germline LOH ref variants  \cr
\strong{n_germ_loh_ref_cov}: Number of germline LOH ref variants covered \cr
\strong{prop_germ_loh_ref_cov}: Number of germline LOH ref variants covered \cr
\strong{alt_ref_ratio}: LOH alt to LOH ref germline variant ratio \cr
\strong{alt_ref_cov_ratio}: LOH alt to LOH ref germline variant coverage ratio \cr
\strong{n_germ_dip}: Number of germline diploid variants \cr
\strong{n_germ_dip_cov}: Number of germline diploid variants covered \cr
\strong{prop_germ_dip_cov}: Proportion of germline diploid variants covered \cr
\strong{mean_dna_cov}: Mean variant coverage in DNA \cr
\strong{med_dna_cov}: Median variant coverage in DNA \cr
\strong{prop_dna_cov}: Proprotion of variants covered in DNA \cr
\strong{n_var_dna_cov}: Number of variants covered in DNA \cr
\strong{mean_rna_cov}: Mean variant coverage in RNA \cr
\strong{med_rna_cov}: Median variant coverage in RNA \cr
\strong{prop_rna_cov}: Proprotion of variants covered in RNA \cr
\strong{n_var_rna_cov}: Number of variants covered in RNA \cr
}

